story update: textarea
